Remove 'not found in _includes' hacks
Remove the use if "unless contains 'not found in _includes directory'"
hacks that were previously used to detect when _includes were missing.

This change is required as the hack no longer works with the latest
Jekyll release (and therefore also with GitHub).

Additional 'custom_pom_template' and 'custom_gradle_template' flags
must now be set in _config.yml if you want to include a custom pom.xml
or build.gradle file.